server/pyproject.toml
[tool.poetry]
name = "cifra-k-test"
version = "0.1.0"
description = ""
authors = ["meldxkviel <mega1hack52@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.3"
dj-database-url = "^1.0.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
